+++ face/__init__.py
@@ -18,7 +18,7 @@ from face.errors import (FaceException,
                          UsageError)
 
 from face.parser import (ListParam, ChoicesParam)
-from face.command import Command
+from face.command import Command, CommandGroup
 from face.middleware import face_middleware
 from face.helpers import HelpHandler, StoutHelpFormatter
 from face.testing import CommandChecker, CheckError