+++ face/parser.py
@@ -560,9 +560,10 @@ class Parser:
     :meth:`Parser.parse()` with ``sys.argv`` or any other list of strings.
     """
     def __init__(self, name, doc=None, flags=None, posargs=None,
-                 post_posargs=None, flagfile=True):
+                 post_posargs=None, flagfile=True, group=None):
         self.name = process_command_name(name)
         self.doc = doc
+        self.group = group
         flags = list(flags or [])
 
         self.posargs = _ensure_posargspec(posargs, 'posargs')